refactor: move workspace mount support
